fix: videoId field type is long.

+++ video-recommendation-service/src/main/java/com/codecool/videorecommendationservice/model/Recommendation.java
@@ -24,6 +24,6 @@ public class Recommendation {
 
     private String comment;
 
-    private int videoId;
+    private long videoId;
 
 }
